refactor: Thread.jsをファイル分割した
